Refactor CI/CD workflows and consolidate in one file
This commit replaces the `cicd-prod.yml` workflow with updates in `cicd.yml`. It factors out the environment specifics through environment variables, allowing the same workflow to be used for both production and development. This makes the workflows more maintainable and ensures correct Docker image is used according to the environment.